Rename all dart files according to the convension: from lowerCamelCase to snake_case.
